refactor: 프로덕션 코드와 테스트 코드 상수 분리
