+++ __tests__/domain/RaceController.test.js
@@ -1,7 +1,7 @@
 import Car from '../../src/domain/Car';
 import RaceCalculator from '../../src/domain/RaceCalculator';
 import Random from '../../src/utils/Random';
-import { RANDOMCASES, TEST_RULES } from '../../src/statics/test_constants';
+import { TEST_RULES } from '../constants/test_constants';
 
 const mockRandoms = numbers => {
   Random.pickNumberInRange = jest.fn();
