+++ docs/REFACTOR.md
@@ -1,4 +1,6 @@
-## 리팩토링 목록
+# 리팩토링 목록
+
+## 1단계
 
 - [x] 상수화
 - [x] 함수 이름 변경
@@ -6,9 +8,15 @@
 - [x] Cars 클래스 존속 여부
 - [x] 도메인 클래스와 유효성 검사 로직 모듈로 분리
 
-## 1차 피드백
+### 1차 피드백
 - [x] 테스트코드 상수화
 - [ ] Car클래스 move 책임 분리
 - [x] RaceService 이름 교체
 - [x] RaceController 파일명 고민 -> RaceCalculator로 변경, domain 폴더 하위로 이동
 - [ ] messages 
+
+
+## 2단계
+- [x] 테스트 코드 상수와 소스 코드 상수 분리 및 상수 분류 확실히
+- [ ] 함수 호출 조건을 인자로 넣어주지 않고 if문을 이용해 가독성을 명확히 하기
+- [ ] 개행 상수 분리하여 사용처에서 조합해 사용토록
